Move user resolvers into dedicated modules
